+++ src/heap.js
@@ -1,6 +1,6 @@
 /**
- * datastructures-js/heap
- * @copyright 2019 Eyas Ranjous <eyas.ranjous@gmail.com>
+ * @datastructures-js/heap
+ * @copyright 2020 Eyas Ranjous <eyas.ranjous@gmail.com>
  * @license MIT
  */
 
@@ -16,56 +16,48 @@ const isNoneEmptyArray = (a) => Array.isArray(a) && a.length > 0;
  * @abstract
  */
 class Heap {
-  constructor() {
-    this.init();
-    this.leafNode = null;
+  constructor(nodes) {
+    this._nodes = Array.isArray(nodes) ? nodes : [];
+    this._leaf = null;
   }
 
   /**
-   * initializes the heap nodes list
-   * @internal
-   */
-  init(nodes) {
-    this.nodes = nodes || [];
-  }
-
-  /**
-   * calculates the left child's index of a parent's index
    * @private
+   * calculates the left child's index of a parent's index
    * @param {number} parentIndex
    * @returns {number}
    */
-  getLeftChildIndex(parentIndex) {
+  _getLeftChildIndex(parentIndex) {
     return (parentIndex * 2) + 1;
   }
 
   /**
-   * calculates the right child's index of a parent's index
    * @private
+   * calculates the right child's index of a parent's index
    * @param {number} parentIndex
    * @returns {number}
    */
-  getRightChildIndex(parentIndex) {
+  _getRightChildIndex(parentIndex) {
     return (parentIndex * 2) + 2;
   }
 
   /**
-   * calculates a parent's index from a child's index
    * @private
+   * calculates a parent's index from a child's index
    * @param {number} parentIndex
    * @returns {number}
    */
-  getParentIndex(childIndex) {
+  _getParentIndex(childIndex) {
     return Math.floor((childIndex - 1) / 2);
   }
 
   /**
-   * gets the last node's index
    * @private
+   * gets the last node's index
    * @returns {number}
    */
-  getLastIndex() {
-    return this.nodes.length - 1;
+  _getLastIndex() {
+    return this._nodes.length - 1;
   }
 
   /**
@@ -73,165 +65,162 @@ class Heap {
    * @param {number} i
    * @param {number} j
    */
-  swap(i, j) {
-    const temp = this.nodes[i];
-    this.nodes[i] = this.nodes[j];
-    this.nodes[j] = temp;
+  _swap(i, j) {
+    const temp = this._nodes[i];
+    this._nodes[i] = this._nodes[j];
+    this._nodes[j] = temp;
   }
 
   /**
-   * returns the correct child's index to fix the heap
    * @private
+   * selects the proper child's index to fix the heap
    * @param {number} parentIndex
    * @returns {number}
    */
-  compareChildrenOf(parentIndex) {
-    const leftChildIndex = this.getLeftChildIndex(parentIndex);
-    const rightChildIndex = this.getRightChildIndex(parentIndex);
+  _compareChildrenOf(parentIndex) {
+    const leftChildIndex = this._getLeftChildIndex(parentIndex);
+    const rightChildIndex = this._getRightChildIndex(parentIndex);
     const size = this.size();
 
-    if (leftChildIndex >= size && rightChildIndex >= size) return 0;
+    if (leftChildIndex >= size && rightChildIndex >= size) return -1;
     if (leftChildIndex >= size) return rightChildIndex;
     if (rightChildIndex >= size) return leftChildIndex;
 
-    return this.compareChildren(leftChildIndex, rightChildIndex);
+    return this._compareChildren(leftChildIndex, rightChildIndex);
   }
 
   /**
+   * @private
    * bubbles the last inserted node up in the heap
-   * @private
    */
-  heapifyUp() {
-    let childIndex = this.getLastIndex();
-    let parentIndex = this.getParentIndex(childIndex);
+  _heapifyUp() {
+    let childIndex = this._getLastIndex();
+    let parentIndex = this._getParentIndex(childIndex);
 
-    while (childIndex > 0 && this.shouldSwap(childIndex, parentIndex)) {
-      this.swap(childIndex, parentIndex);
+    while (this._shouldSwap(childIndex, parentIndex)) {
+      this._swap(childIndex, parentIndex);
       childIndex = parentIndex;
-      parentIndex = this.getParentIndex(childIndex);
+      parentIndex = this._getParentIndex(childIndex);
     }
   }
 
   /**
-   * pushes the replaced root node down in the heap after root's removal
    * @private
+   * pushes the replaced root node down in the heap after root's removal
    */
-  heapifyDown() {
+  _heapifyDown() {
     let parentIndex = 0;
-    let childIndex = this.compareChildrenOf(parentIndex);
-
-    while (childIndex > 0 && this.shouldSwap(childIndex, parentIndex)) {
-      this.swap(childIndex, parentIndex);
+    let childIndex = this._compareChildrenOf(parentIndex);
+    while (this._shouldSwap(childIndex, parentIndex)) {
+      this._swap(childIndex, parentIndex);
       parentIndex = childIndex;
-      childIndex = this.compareChildrenOf(parentIndex);
+      childIndex = this._compareChildrenOf(parentIndex);
     }
   }
 
   /**
+   * @private
    * pushes the swapped node with root down in its correct location
    * @param {number} i -  swapped node's index
-   * @private
    */
-  heapifyDownUntil(index) {
+  _heapifyDownUntil(index) {
     let parentIndex = 0;
     let leftChildIndex = 1;
     let rightChildIndex = 2;
     let childIndex;
 
     while (leftChildIndex < index) {
-      childIndex = this.compareChildrenBefore(
+      childIndex = this._compareChildrenBefore(
         index,
         leftChildIndex,
         rightChildIndex
       );
 
-      if (this.shouldSwap(childIndex, parentIndex)) {
-        this.swap(childIndex, parentIndex);
+      if (this._shouldSwap(childIndex, parentIndex)) {
+        this._swap(childIndex, parentIndex);
       }
 
       parentIndex = childIndex;
-      leftChildIndex = this.getLeftChildIndex(parentIndex);
-      rightChildIndex = this.getRightChildIndex(parentIndex);
+      leftChildIndex = this._getLeftChildIndex(parentIndex);
+      rightChildIndex = this._getRightChildIndex(parentIndex);
     }
   }
 
   /**
-   * implements heap sort algorithm by swapping root with i nodes
-   * @public
-   * @returns {array} sorted nodes list
-   */
-  sort() {
-    for (let i = this.getLastIndex(); i > 0; i -= 1) {
-      this.swap(0, i);
-      this.heapifyDownUntil(i);
-    }
-
-    return this.nodes;
-  }
-
-  /**
-   * returns a shallow copy of a heap
    * @protected
+   * returns a shallow copy of a heap
    * @param {class} HeapType
    * @returns {Heap}
    */
-  clone(HeapType) {
-    const heap = new HeapType();
-    heap.init(this.nodes.slice());
-    return heap;
+  _clone(HeapType) {
+    return new HeapType(this._nodes.slice());
   }
 
   /**
+   * @public
+   * implements heap sort algorithm by swapping root with i nodes
+   * @returns {array} the sorted nodes
+   */
+  sort() {
+    for (let i = this._getLastIndex(); i > 0; i -= 1) {
+      this._swap(0, i);
+      this._heapifyDownUntil(i);
+    }
+
+    return this._nodes;
+  }
+
+  /**
+   * @public
    * inserts a node into the heap
    * @param {number|string} key
    * @param {object} value
    * @returns {HeapNode}
-   * @public
    */
   insert(key, value) {
     const newNode = new HeapNode(key, value);
-    this.nodes.push(newNode);
-    this.heapifyUp();
+    this._nodes.push(newNode);
+    this._heapifyUp();
     return newNode;
   }
 
   /**
-   * returns the root node in the heap
    * @public
+   * returns the root node in the heap
    * @returns {HeapNode}
    */
   root() {
-    if (this.size() === 0) return null;
-    return this.nodes[0];
+    if (this.isEmpty()) return null;
+    return this._nodes[0];
   }
 
   /**
-   * returns a leaf node in the heap
    * @public
+   * returns a leaf node in the heap
    * @returns {HeapNode}
    */
   leaf() {
-    return this.leafNode;
+    return this._leaf;
   }
 
   /**
-   * removes and returns the root node in the heap
    * @public
+   * removes and returns the root node in the heap
    * @returns {HeapNode}
    */
   extractRoot() {
-    if (this.size() === 0) return null;
+    if (this.isEmpty()) return null;
 
-    const root = this.nodes[0];
-    this.nodes[0] = this.nodes[this.getLastIndex()];
-    this.nodes.pop();
-    this.heapifyDown();
+    const root = this.root();
+    this._nodes[0] = this._nodes[this._getLastIndex()];
+    this._nodes.pop();
+    this._heapifyDown();
 
-    if (root === this.leafNode) {
-      if (this.size() === 0) {
-        this.leafNode = null;
+    if (root === this._leaf) {
+      if (this.isEmpty()) {
+        this._leaf = null;
       } else {
-        this.leafNode = this.root();
+        this._leaf = this.root();
       }
     }
 
@@ -239,40 +228,41 @@ class Heap {
   }
 
   /**
-   * returns the number of nodes in the heap
    * @public
+   * returns the number of nodes in the heap
    * @returns {number}
    */
   size() {
-    return this.nodes.length;
+    return this._nodes.length;
   }
 
   /**
-   * serializes the heap nodes into a list of objects
    * @public
-   * @returns {array}
+   * checks if the heap is empty
+   * @returns {boolean}
    */
-  serialize() {
-    return this.nodes.map((n) => n.serialize());
+  isEmpty() {
+    return this.size() === 0;
   }
 
   /**
-   * clears the heap
    * @public
+   * clears the heap
    */
   clear() {
-    this.init();
+    this._nodes = [];
+    this._leaf = null;
   }
 
   /**
+   * @protected
+   * @static
    * convert a list of items into a heap
    * @param {array} items
    * @param {class} HeapType
-   * @protected
-   * @static
    * @returns {Heap}
    */
-  static heapify(items, HeapType) {
+  static _heapify(items, HeapType) {
     if (!isNoneEmptyArray(items)) return null;
 
     const heap = new HeapType();