+++ src/minHeap.js
@@ -12,40 +12,15 @@ const Heap = require('./heap');
  */
 class MinHeap extends Heap {
   /**
-   * inserts a node into the heap and rebase leaf node to max key
-   * @param {number|string} key
-   * @param {object} value
-   * @public
-   * @override
-   */
-  insert(key, value) {
-    const newNode = super.insert(key, value);
-    if (this.leafNode === null || key > this.leafNode.getKey()) {
-      this.leafNode = newNode;
-    }
-  }
-
-  /**
-   * checks if child's key is smaller than its parent's key
-   * @protected
-   * @returns {boolean}
-   */
-  shouldSwap(childIndex, parentIndex) {
-    const child = this.nodes[childIndex];
-    const parent = this.nodes[parentIndex];
-    return child.getKey() < parent.getKey();
-  }
-
-  /**
-   * gets the min child's index of two node's children
    * @private
+   * gets the min child's index of two node's children
    * @param {number} leftChildIndex
    * @param {number} rightChildIndex
    * @returns {number}
    */
-  getMinChildIndex(leftChildIndex, rightChildIndex) {
-    const leftChild = this.nodes[leftChildIndex];
-    const rightChild = this.nodes[rightChildIndex];
+  _getMinChildIndex(leftChildIndex, rightChildIndex) {
+    const leftChild = this._nodes[leftChildIndex];
+    const rightChild = this._nodes[rightChildIndex];
     if (leftChild.getKey() < rightChild.getKey()) {
       return leftChildIndex;
     }
@@ -53,16 +28,16 @@ class MinHeap extends Heap {
   }
 
   /**
-   * returns the min child's index of two children before an index
    * @private
+   * returns the min child's index of two children before an index
    * @param {number} index
    * @param {number} leftChildIndex
    * @param {number} rightChildIndex
    * @returns {number}
    */
-  getMinChildIndexBefore(index, leftChildIndex, rightChildIndex) {
-    const leftChild = this.nodes[leftChildIndex];
-    const rightChild = this.nodes[rightChildIndex];
+  _getMinChildIndexBefore(index, leftChildIndex, rightChildIndex) {
+    const leftChild = this._nodes[leftChildIndex];
+    const rightChild = this._nodes[rightChildIndex];
     if (rightChild.getKey() < leftChild.getKey() && rightChildIndex < index) {
       return rightChildIndex;
     }
@@ -70,50 +45,77 @@ class MinHeap extends Heap {
   }
 
   /**
-   * implements the parent's function to select a child's index
+   * @protected
+   * checks if child's key is smaller than its parent's key
+   * @returns {boolean}
+   */
+  _shouldSwap(childIndex, parentIndex) {
+    if (childIndex < 0 || childIndex >= this.size()) return false;
+    if (parentIndex < 0 || parentIndex >= this.size()) return false;
+
+    const child = this._nodes[childIndex];
+    const parent = this._nodes[parentIndex];
+
+    return child.getKey() < parent.getKey();
+  }
+
+  /**
    * @protected
    * @override
+   * implements the parent's function to select a child's index
    * @param {number} leftChildIndex
    * @param {number} rightChildIndex
    * @returns {number}
    */
-  compareChildren(leftChildIndex, rightChildIndex) {
-    return this.getMinChildIndex(leftChildIndex, rightChildIndex);
+  _compareChildren(leftChildIndex, rightChildIndex) {
+    return this._getMinChildIndex(leftChildIndex, rightChildIndex);
   }
 
   /**
-   * implements the parent's function to select a child's index before an index
    * @protected
    * @override
+   * implements the parent's function to select a child's index before an index
    * @param {number} index
    * @param {number} leftChildIndex
    * @param {number} rightChildIndex
    * @returns {number}
    */
-  compareChildrenBefore(index, leftChildIndex, rightChildIndex) {
-    return this.getMinChildIndexBefore(index, leftChildIndex, rightChildIndex);
+  _compareChildrenBefore(index, leftChildIndex, rightChildIndex) {
+    return this._getMinChildIndexBefore(index, leftChildIndex, rightChildIndex);
   }
 
   /**
-   * returns a shallow copy of a min heap
    * @public
    * @override
+   * inserts a node into the heap and rebase leaf node to max key
+   * @param {number|string} key
+   * @param {object} value
+   */
+  insert(key, value) {
+    const newNode = super.insert(key, value);
+    if (this._leaf === null || key > this._leaf.getKey()) {
+      this._leaf = newNode;
+    }
+    return newNode;
+  }
+
+  /**
+   * @public
+   * returns a shallow copy of a min heap
    * @returns {MinHeap}
    */
   clone() {
-    return super.clone(MinHeap);
+    return super._clone(MinHeap);
   }
 
   /**
+   * @public
    * builds a min heap from an array of items
    * @param {array} items
-   * @public
-   * @override
-   * @static
    * @returns {MinHeap}
    */
   static heapify(items) {
-    return super.heapify(items, MinHeap);
+    return super._heapify(items, MinHeap);
   }
 }